backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlmodel>=0.0.14
pyjwt[crypto]>=2.8.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.5
passlib[bcrypt]>=1.7.4
jwcrypto>=1.5.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
httpx>=0.26.0
openai>=1.3.0
email-validator>=2.0.0   
psycopg2-binary
